Updated Alternitive
This is my updated alternative for this game. As I was not able to implement any more code to our hue game (due to lack of understanding from my own part), I made this to feel less anxious on our submission. Please give it a chance and run it. I believe it meets the requirements. However, It is up to you guys if you were able to make changes to the hue game and would like to submit that instead.